Annotation update
Updating intrinsics to include cpuid code
Updating suspect list to include cpuid, eflags, and unresolved register

+++ annotations/nocedit.exe/pseudocode/src/core/dstrender.cpp/blendBilerpLightmapToTexture64BitOutput_FUN_00491c9a.cpp
@@ -39,10 +39,11 @@ core_dstrender_cpp_blendBilerpLightmapToTexture64BitOutput_FUN_00491c9a
   uint5 uVar25;
   
   uVar24 = psllw((ulonglong)
-                 CONCAT52(CONCAT41((int)(((uint7)CONCAT21((short)(((uint7)(byte)((uint)
-                                                  g_SolidColorMode >> 0x18) << 0x30) >> 0x28),
-                                                  (char)((uint)g_SolidColorMode >> 0x10)) << 0x20)
-                                        >> 0x18),(char)((uint)g_SolidColorMode >> 8)),
+                 CONCAT52(CONCAT41((int)(((uint7)CONCAT21((short)(((uint7)(byte)(g_SolidColorMode >>
+                                                                                0x18) << 0x30) >>
+                                                                 0x28),
+                                                          (char)(g_SolidColorMode >> 0x10)) << 0x20)
+                                        >> 0x18),(char)(g_SolidColorMode >> 8)),
                           (ushort)(byte)g_SolidColorMode),6);
   do {
     uVar1 = *texture_buffer;
