+++ annotations/nocedit.exe/pseudocode/src/core/dstrender.cpp/blendLightmapToTextureMMX_FUN_00493dc4.cpp
@@ -28,10 +28,11 @@ core_dstrender_cpp_blendLightmapToTextureMMX_FUN_00493dc4
   ulonglong uVar14;
   
   uVar14 = psllw((ulonglong)
-                 CONCAT52(CONCAT41((int)(((uint7)CONCAT21((short)(((uint7)(byte)((uint)
-                                                  g_SolidColorMode >> 0x18) << 0x30) >> 0x28),
-                                                  (char)((uint)g_SolidColorMode >> 0x10)) << 0x20)
-                                        >> 0x18),(char)((uint)g_SolidColorMode >> 8)),
+                 CONCAT52(CONCAT41((int)(((uint7)CONCAT21((short)(((uint7)(byte)(g_SolidColorMode >>
+                                                                                0x18) << 0x30) >>
+                                                                 0x28),
+                                                          (char)(g_SolidColorMode >> 0x10)) << 0x20)
+                                        >> 0x18),(char)(g_SolidColorMode >> 8)),
                           (ushort)(byte)g_SolidColorMode),6);
   do {
     uVar7 = *texture_buffer;
