+++ annotations/nocedit.exe/pseudocode/src/core/dstrender.cpp/blendSinglePixelLightmapMMX_FUN_004937b6.cpp
@@ -26,10 +26,11 @@ core_dstrender_cpp_blendSinglePixelLightmapMMX_FUN_004937b6
   ulonglong uVar13;
   
   uVar13 = psllw((ulonglong)
-                 CONCAT52(CONCAT41((int)(((uint7)CONCAT21((short)(((uint7)(byte)((uint)
-                                                  g_SolidColorMode >> 0x18) << 0x30) >> 0x28),
-                                                  (char)((uint)g_SolidColorMode >> 0x10)) << 0x20)
-                                        >> 0x18),(char)((uint)g_SolidColorMode >> 8)),
+                 CONCAT52(CONCAT41((int)(((uint7)CONCAT21((short)(((uint7)(byte)(g_SolidColorMode >>
+                                                                                0x18) << 0x30) >>
+                                                                 0x28),
+                                                          (char)(g_SolidColorMode >> 0x10)) << 0x20)
+                                        >> 0x18),(char)(g_SolidColorMode >> 8)),
                           (ushort)(byte)g_SolidColorMode),6);
   uVar1 = *texture_pixel;
   uVar2 = g_LightmapTexturePalette[*texture_index];
